+++ lib/apis/core/pairing/pairing.dart
@@ -78,6 +78,7 @@ class Pairing implements IPairing {
 
     _registerRelayEvents();
     _registerExpirerEvents();
+    _registerheartbeatSubscription();
 
     await core.expirer.init();
     await pairings.init();
@@ -92,9 +93,14 @@ class Pairing implements IPairing {
   }
 
   @override
-  Future<CreateResponse> create({
-    List<List<String>>? methods,
-  }) async {
+  Future<void> checkAndExpire() async {
+    for (var pairing in getPairings()) {
+      await core.expirer.checkAndExpire(pairing.topic);
+    }
+  }
+
+  @override
+  Future<CreateResponse> create({List<List<String>>? methods}) async {
     _checkInitialized();
     final String symKey = core.crypto.getUtils().generateRandomBytes32();
     final String topic = await core.crypto.setSymKey(symKey);
@@ -233,7 +239,7 @@ class Pairing implements IPairing {
     required ProtocolType type,
   }) {
     if (routerMapRequest.containsKey(method)) {
-      throw WalletConnectError(
+      throw const WalletConnectError(
         code: -1,
         message: 'Method already exists',
       );
@@ -277,7 +283,7 @@ class Pairing implements IPairing {
         WalletConnectUtils.calculateExpiry(
           WalletConnectConstants.THIRTY_DAYS,
         )) {
-      throw WalletConnectError(
+      throw const WalletConnectError(
         code: -1,
         message: 'Expiry cannot be more than 30 days away',
       );
@@ -308,6 +314,11 @@ class Pairing implements IPairing {
   @override
   List<PairingInfo> getPairings() {
     return pairings.getAll();
+  }
+
+  @override
+  PairingInfo? getPairing({required String topic}) {
+    return pairings.get(topic);
   }
 
   @override
@@ -385,7 +396,7 @@ class Pairing implements IPairing {
     int? ttl,
     EncodeOptions? encodeOptions,
   }) async {
-    core.logger.v(
+    core.logger.t(
       'pairing sendResult, id: $id topic: $topic, method: $method, params: $params, ttl: $ttl',
     );
 
@@ -456,7 +467,7 @@ class Pairing implements IPairing {
     dynamic result, {
     EncodeOptions? encodeOptions,
   }) async {
-    core.logger.v(
+    core.logger.t(
       'pairing sendResult, id: $id topic: $topic, method: $method, result: $result',
     );
     final Map<String, dynamic> payload =
@@ -491,7 +502,7 @@ class Pairing implements IPairing {
     JsonRpcError error, {
     EncodeOptions? encodeOptions,
   }) async {
-    core.logger.v(
+    core.logger.t(
       'pairing sendError, id: $id topic: $topic, method: $method, error: $error',
     );
 
@@ -524,6 +535,11 @@ class Pairing implements IPairing {
   /// ---- Private Helpers ---- ///
 
   Future<void> _resubscribeAll() async {
+    // If the relay is not active, stop here
+    if (!core.relayClient.isConnected) {
+      return;
+    }
+
     // Resubscribe to all active pairings
     final List<PairingInfo> activePairings = pairings.getAll();
     for (final PairingInfo pairing in activePairings) {
@@ -639,6 +655,7 @@ class Pairing implements IPairing {
     // print(payloadString);
 
     Map<String, dynamic> data = jsonDecode(payloadString);
+    core.logger.i('Pairing _onMessageEvent, Received data: $data');
 
     // If it's an rpc request, handle it
     // print('Pairing: Received data: $data');
@@ -774,6 +791,10 @@ class Pairing implements IPairing {
     core.expirer.onExpire.subscribe(_onExpired);
   }
 
+  void _registerheartbeatSubscription() {
+    core.heartbeat.onPulse.subscribe(_heartbeatSubscription);
+  }
+
   Future<void> _onExpired(ExpirationEvent? event) async {
     if (event == null) {
       return;
@@ -788,6 +809,11 @@ class Pairing implements IPairing {
         ),
       );
     }
+  }
+
+  void _heartbeatSubscription(EventArgs? args) async {
+    core.logger.i('Pairing heartbeat received');
+    await checkAndExpire();
   }
 
   /// ---- Validators ---- ///