+++ lib/apis/utils/constants.dart
@@ -1,11 +1,12 @@
 class WalletConnectConstants {
-  static const SDK_VERSION = '2.0.9';
+  static const SDK_VERSION = '2.1.8';
 
   static const CORE_PROTOCOL = 'wc';
   static const CORE_VERSION = 2;
   static const CORE_CONTEXT = 'core';
 
-  static const DEFAULT_RELAY_URL = 'wss://relay.walletconnect.org';
+  static const DEFAULT_RELAY_URL = 'wss://relay.walletconnect.com';
+  static const FALLBACK_RELAY_URL = 'wss://relay.walletconnect.org';
 
   static const CORE_STORAGE_PREFIX =
       '$CORE_PROTOCOL@$CORE_VERSION:$CORE_CONTEXT:';