+++ test/core_api/store_test.dart
@@ -208,7 +208,7 @@ void main() {
 
         await specialStore.set(
           '1',
-          JsonRpcRecord(
+          const JsonRpcRecord(
             id: 1,
             topic: 'test',
             method: 'method',